+++ cli/cli/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+- config files use indented JSON
+
 ## [1.19.17] - 2024-04-04
 ### Changed
 - `BEAM_DOCKER_URI` environment variable will override docker connection uri